finished booking cards, started on venuemanager section

+++ src/Utilities/DateFormat/index.jsx
@@ -1,6 +1,6 @@
 function formatDate(date) {
   const dateObj = new Date(date);
-  const options = { day: "numeric", month: "long", year: "numeric" };
+  const options = { day: "numeric", month: "numeric", year: "numeric" };
   return dateObj.toLocaleDateString(undefined, options);
 }
